requirements.txt
certifi==2019.3.9
chardet==3.0.4
defusedxml==0.5.0
Django==2.1.7
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap4==0.0.7
django-chartjs==1.3
django-extensions==2.1.5
idna==2.8
oauthlib==3.0.1
PyJWT==1.7.1
python3-openid==3.1.0
pytz==2018.9
requests==2.21.0
requests-oauthlib==1.2.0
six==1.12.0
social-auth-app-django==3.1.0
social-auth-core==3.1.0
urllib3==1.24.1
